requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dotenv==0.9.9
frozenlist==1.8.0
gazpacho==1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
multidict==6.7.0
packaging==26.0
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
six==1.17.0
soupsieve==2.8.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.3.0
yarl==1.22.0
